Insert the first version of the source list, unsorted

+++ Objective C/Séries/RakSRSearchList.h
@@ -14,7 +14,7 @@
 {
 	byte _type;
 	
-	uint * indexes;
+	uint64_t * indexes;
 	
 	BOOL lastWasSelection;
 	NSMutableArray * selection;
